refactor openVPN management
