+++ OpenVPNGateMonitor/Program.cs
@@ -2,7 +2,7 @@ using OpenVPNGateMonitor.Configurations;
 
 var builder = WebApplication.CreateBuilder(args);
 
-builder.Services.ConfigureServices();
+builder.Services.ConfigureServices(builder.Configuration);
 builder.Services.DataBaseServices(builder.Configuration);
 
 builder.Host.ConfigureSerilog(builder.Configuration);
