+++ OpenVPNGateMonitor/Services/OpenVpnManagementInterfaces/Interfaces/IGeoIpService.cs
@@ -1,0 +1,8 @@
+﻿using OpenVPNGateMonitor.Models.Helpers.OpenVpnManagementInterfaces;
+
+namespace OpenVPNGateMonitor.Services.OpenVpnManagementInterfaces.Interfaces;
+
+public interface IGeoIpService
+{
+    OpenVpnGeoInfo? GetGeoInfo(string ip);
+}
