+++ OpenVPNGateMonitor/Services/OpenVpnManagementInterfaces/OpenVpnVersionService.cs
@@ -1,18 +1,24 @@
-﻿using OpenVPNGateMonitor.Services.BackgroundServices.Interfaces;
+﻿using MaxMind.GeoIP2;
+using OpenVPNGateMonitor.Services.BackgroundServices.Interfaces;
 using OpenVPNGateMonitor.Services.OpenVpnManagementInterfaces.Interfaces;
 
 namespace OpenVPNGateMonitor.Services.OpenVpnManagementInterfaces;
 
 public class OpenVpnVersionService
 {
-    public OpenVpnVersionService(IOpenVpnManagementService)
+    private readonly ILogger<OpenVpnVersionService> _logger;
+    private readonly IOpenVpnManagementService _openVpnManagementService;
+    
+    public OpenVpnVersionService(ILogger<OpenVpnVersionService> logger, 
+        IOpenVpnManagementService openVpnManagementService)
     {
-        
+        _logger = logger;
+        _openVpnManagementService = openVpnManagementService;
     }
     
     public async Task<string> GetVersionAsync(CancellationToken cancellationToken)
     {
-        return await SendCommandAsync("version", cancellationToken);
+        return await _openVpnManagementService.SendCommandAsync("version", cancellationToken);
     }
 
 }
